Implement favourites tab for clocks
Added a star button that lets the user favourite a timer or
a stopwatch to add it to the new favourites tab, giving it
precedence over other clocks. If a user has 3 timers and 5 stopwatches
and presses the star button on the 4th stopwatch it will now appear
at the top of the clocks tab over any clock that isn't favourited.

Co-authored-by: Francisco Alves francisco.p.alves@tecnico.ulisboa.pt